KTIJ-31412 Extract SuccessfulImportFeature from NoErrorEventsDuringImportFeature
Some old import tests (e.g. where a KGP version is 1.7.21) finished with `BUILD FAILED` state from the Gradle daemon, but import itself is robust enough to extract all necessary information correctly. The original purpose of the NoErrorEventsDuringImportFeature was to collect import errors, but the `BUILD FAILED` state is not one of them. Also, this feature is enabled by default for all tests, while we need to check that the Gradle build was successful only for the Isolation Project feature tests

GitOrigin-RevId: 927457e150beefbee2c7f9ac8fd73e5255bd6b10