[kotlin] KTIJ-35659 Fix escaping for pasted external strings
In K2 mode, always escape text parsed as interpolation entry
when pasting from outside the IDEA into a single-quoted string.
Pasting between strings inside the IDEA is handled by the post-processor,
so the entries remain entries for the non-outsider strings.
In K1 the old non-escaping behavior is necessary because there is no post-processor.

GitOrigin-RevId: 1ae957d90de6cc5fbf0e96a7d4854c979e7b5f18