[kotlin] KTIJ-35825 Avoid traversing object's `memberScope`s when there are no potential callables to be inherited
This should greatly improve the speed with which the auto-import hint and quickfix are computed
for the cases when the unresolved name is rare enough not to appear in too many places.

At the same time, if some unresolved reference can point to a declaration inherited by an object,
then finding a base symbol for it should be fast enough, and the `none()` check should return `false`
quickly.

Add a few tests to explicitly check that methods and fields from Java are still
properly inherited.

^KTIJ-35825 Fixed

GitOrigin-RevId: 3e1dad437ff2a5a48c59c1bb63212aa87402af2a