[kotlin] KT-77708 Fix redundant K2 smart cast highlighting with expected type
K1 compiler performs the smart cast only when the original type
is different from the expected type. K2 compiler, however, always
performs an eager smart cast regardless of the expected type.

This difference in behavior is expected and won't be changed.

To fix the redundant highlighting on the IDE side, we use the new API
`KaSmartCastInfo.originalType`. The smart cast highlighting
is considered redundant if the original expression type
is compatible with its expected type.

^KT-77708 Fixed

GitOrigin-RevId: 86beeebb24665a7a7aaec2a58c86f5ae5537dae6