[kotlin] KTIJ-30268 K2 Move: fix EDT analysis exception when moving files
EDT analysis should be allowed explicitly as it can cause freezes when
used in an inappropriate context. The move refactoring goes under a
modal progress indicator, so using resolve is not an issue. Move requires
resolving references. For regular references the lack of explicit
permission didn't hurt as the Kotlin reference resolver does allow
EDT analysis on its own. However, for references from UAST providers
there was no such permission, which caused exceptions and misfunctioning
refactoring. For declarations move the analysis on EDT is already allowed.

GitOrigin-RevId: 441e9d2033254a0cf0ac6c75835a8060b7c04d4a