[kotlin] KTIJ-2308 Implement action to navigate to conflicting declarations
The implementation of the action itself is closely based on a
similar action for Java: IDEA-216276

It is currently supported for the following three diagnostic:
- ConflictingOverloads for functions
- Redeclaration for properties
- ClassifierRedeclaration for classes

In `AbstractQuickFixTest`, add the code which waits for
the execution of all `NonBlockingReadAction`-s - this is
required to wait until all the mod commands are executed.
This is relevant for this fix, because the `ModCommand.select`
with a single option is executed in two steps, and we want to
wait for them both.

Also, replace `dispatchAllInvocationEvents` with `dispatchAllInvocationEventsInIdeEventQueue`
(as it is recommended in the javadoc of the former)

^KTIJ-2308 Fixed

GitOrigin-RevId: 6b03fecb98f45871b837455f81aaec1474f12dcf