IDEA-379042 Favor 'ClassNameProvider' over 'ClassNameCalculator' & Drop ComposeClassNameCalculator in favor of ComposeClassNameContributor
The ClassNameCalculator and ClassNameProvider provide a very similar-looking API.
The ClassNameProvider even uses the ClassNameCalculator internally.
Conceptionally, however, the ClassNameProvider provides a full list of 'candidates'
for potential 'Class Names' which could contain a given element or source location, whereas
the 'ClassNameCalculator' tries to closely match the output of the vanilla JVM backend.

Since the exact calculation of the ComposablesSingletons requires resolving calls, which
might be too slow for certain use-cases, the Compose integration is done by contributing
'ComposableSingleton' class names as 'candidates' for the 'ClassNameProvider' instead.

GitOrigin-RevId: efc4aeab18eb7d77e0ae14d03304a730fb10f70d