cleanup [kotlin]: remove unused incorrect dependency on intellij.platform.structuralSearch from intellij.kotlin
This dependency isn't used in code, and there is no dependency on 'intellij.platform.structuralSearch' in the XML descriptor, which was converted to a content module, so now it's necessary to specify an explicit dependency on it.

This is needed to fix IdeaUltimatePluginModuleDependenciesTest after restoring checks for dependencies on modules in the core plugin (IJPL-224619).

GitOrigin-RevId: a1ed4540eb33db5d98dd9a96b3c7bf8751976770